Initial commit of tools. Most of the tools are recursive find version of the equivalent tools. For example, git_commit <file_name> is the same as git commit <relative path to file_name>.
new branch commit is a pretty nice interactive script for cherry picking files into different branches and commits. It is interactive and will prompt to view diff, push branch, and cleanup. Input is a new branch name to create and list of files to commit.

git_squash is the last tool worth an honorable mention, can be used to simply squash the last "n" commits together.